requirements.txt
certifi==2019.3.9
chardet==3.0.4
Deprecated==1.2.5
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
plumbum==1.6.7
PyGithub==1.43.7
PyJWT==1.7.1
requests==2.22.0
smmap2==2.0.5
urllib3==1.25.2
wrapt==1.11.1
